requirements.txt
PyQt5>=5.15.0
numpy>=1.21.0
Pillow>=8.3.0
pydicom>=2.3.0
scipy>=1.7.0
scikit-image>=0.18.0
